chore: add memoization
