+++ src/js/index.js
@@ -6,23 +6,7 @@ import { ResultModal, initResultModalParam } from "./view/ResultModal.js";
 import { lottoStore } from './store/LottoStore.js';
 import { resultStore } from './store/ResultStore.js';
 
-function memoization(Component, initParam) {
-  const cache = { ...initParam };
-
-  return (store) => {
-    let isChange = false;
-    for (const key in cache) {
-      if (cache[key] === store[key]) {
-        isChange = true;
-      }
-      cache[key] = store[key];
-    }
-
-    if (isChange) {
-      Component(cache);
-    }
-  }
-}
+import { memoization } from './core/Memoization';
 
 const memoizedLottoList = memoization(LottoList, initParmLottoList);
 const memoizedMyLotto = memoization(MyLotto, initMyLottoParam);
